fix(sptxx): fix ctor_wrapper UB crash and eliminate all warnings
- usertype.hpp ctor_wrapper: initialize T* to nullptr after
  lua_newuserdatauv, so __gc reads nullptr (skips delete) instead of
  garbage pointer when extract_args/construct throws. This was the root
  cause of test_edge_usertype crashing in lua_close with HEAP_CORRUPTION.
- usertype.hpp binary_operator_wrapper: use function_traits to dispatch
  scalar second operand by parameter type (int vs float), eliminating
  C4244 implicit narrowing warnings (lua_Integer to float/double).
- test_call_metamethod.cpp / test_list_bounds.cpp: fix compile errors
  (LUA_TABLIBNAME to LUA_LISTLIBNAME for SPT customized Lua).
- test_edge_table_proxy.cpp / test_usertype_operators.cpp: add
  static_cast to fix C4267/C4244 narrowing warnings.

All 48 sptxx tests pass with zero warnings.

+++ sptxx/tests/test_edge_table_proxy.cpp
@@ -147,7 +147,7 @@ int main() {
     {
       auto p1 = lua["a"]["b"];
       auto p2 = p1; // 拷贝
-      int v = p2["c"]["d"].get<std::string>().size(); // d 现在是 "text"
+      int v = static_cast<int>(p2["c"]["d"].get<std::string>().size()); // d 现在是 "text"
       CHECK(v == 4, "copied proxy still works, len=\"text\" = " << v);
     }
